+++ Flow/Changes.md
@@ -31,6 +31,7 @@ Seq in CollectionExtensions
 - Added .current() to WeekDay
 - Added .popAll() to VolatileList
 - Added .compareBy(...), .compareWith(...), .merge(...) and .mergeBy(...) methods to ChangeEvent
+- Added .onAnyChange(...) to ChangeListener (type)
 
 ### Fixes
 - Fixed a bug in LocalDate.next(WeekDay) and LocalDate.previous(WeekDay) (returned same day even when